# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "description": "Retorna uma lista paginada de produtos associados às credenciais do usuário autenticado.",
        "operationId": "ProductController_getProducts",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "description": "Código da loja para buscar os estoques dos produtos",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página para paginação",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Número de itens por página",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "priceTable",
            "required": false,
            "in": "query",
            "description": "Código da tabela de preço para buscar os preços dos produtos. Se não informado, será usado o valor 1.",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "group",
            "required": false,
            "in": "query",
            "description": "Código do grupo para filtrar produtos",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "brand",
            "required": false,
            "in": "query",
            "description": "Código da marca para filtrar produtos",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minStock",
            "required": false,
            "in": "query",
            "description": "Quantidade mínima em estoque para filtrar produtos",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Termo de busca para filtrar produtos por descrição",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de requisição, como pageSize excedendo o limite."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar produtos",
        "tags": [
          "Product"
        ]
      }
    },
    "/api/v1/products/id/{id}": {
      "get": {
        "description": "Retorna os detalhes de um produto específico com base no ID fornecido.",
        "operationId": "ProductController_getProductById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do produto a ser retornado",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "priceTable",
            "required": false,
            "in": "query",
            "description": "Código da tabela de preço para buscar os preços dos produtos. Se não informado, será usado o valor 1.",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "Código da loja para buscar os estoques dos produtos. Se não informado, será usado o valor 1.",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do produto retornados com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Erro de requisição, como ID inválido."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de um produto pelo ID",
        "tags": [
          "Product"
        ]
      }
    },
    "/api/v1/products/barcode/{barcode}": {
      "get": {
        "description": "Retorna os detalhes de um produto específico com base no código de barras fornecido.",
        "operationId": "ProductController_getProductByBarcode",
        "parameters": [
          {
            "name": "barcode",
            "required": true,
            "in": "path",
            "description": "Código de barras do produto a ser retornado",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "priceTable",
            "required": false,
            "in": "query",
            "description": "Código da tabela de preço para buscar os preços dos produtos. Se não informado, será usado o valor 1.",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "Código da loja para buscar os estoques dos produtos. Se não informado, será usado o valor 1.",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do produto retornados com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductBarcodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Erro de requisição, como código de barras inválido."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de um produto pelo código de barras",
        "tags": [
          "Product"
        ]
      }
    },
...
    "/api/v1/service-providers": {
      "get": {
        "operationId": "ServiceProviderController_get",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página atual",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca por nome ou apelido",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "situation",
            "required": false,
            "in": "query",
            "description": "Situação do prestador",
            "schema": {
              "type": "string",
              "enum": [
                "A",
                "I"
              ]
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "ID da loja",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de prestadores de serviço",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceProviderResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar prestadores de serviço",
        "tags": [
          "Service Providers (Prestadores)"
        ]
      }
    },
    "/api/v1/service-providers/{id}": {
      "get": {
        "operationId": "ServiceProviderController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do prestador (PRE_CODIGO)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do prestador de serviço",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProviderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhe do prestador",
        "tags": [
          "Service Providers (Prestadores)"
        ]
      }
    },
...
  },
  "info": {
    "title": "InfoBrasil API",
    "description": "API de integração do sistema Retaguarda (by InfoBrasil) com outros sistemas, como ERPs, e-commerce, etc. <br><br>\n      Se deseja solicitar um acesso, preencha o formulário em: <a href='https://info-api.infobrasilsistemas.com.br/integration/form'>Solicitar Acesso</a>\n      <br><br>\n      **O sistema conta com duas formas de autenticação:**<br> \n      - Basic Auth para login e obtenção de token JWT<br>\n      - JWT para acesso aos endpoints protegidos.\n      <br><br>\n      Para verificar a disponibilidade da API acesse nossa <a href='/status'>Página de Status</a>\n      <br><br>\n      A API é organizada em módulos, cada um responsável por uma área específica do sistema, como produtos, clientes, vendas, etc.",
    "version": "1.15.1",
    "contact": {
      "name": "InfoBrasil Sistemas",
      "url": "https://www.infobrasilsistemas.com.br",
      "email": "suporte@infobrasilsistemas.com.br"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3336",
      "description": "Servidor local para desenvolvimento"
    },
    {
      "url": "https://info-api.infobrasilsistemas.com.br",
      "description": "Servidor de produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "PRO_CODIGO": {
            "type": "number",
            "example": 101,
            "description": "Código do produto"
          },
          "PRO_CODIGOBAR": {
            "type": "string",
            "example": "7891234567890",
            "description": "Código de barras"
          },
          "PRO_DESCRICAO": {
            "type": "string",
            "example": "Produto Exemplo",
            "description": "Descrição do produto"
          },
          "MAR_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da marca"
          },
          "MAR_DESCRICAO": {
            "type": "string",
            "example": "Marca Exemplo",
            "description": "Descrição da marca"
          },
          "GRU_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do grupo"
          },
          "GRU_DESCRICAO": {
            "type": "string",
            "example": "Grupo Exemplo",
            "description": "Descrição do grupo"
          },
          "EST_ATUAL": {
            "type": "number",
            "example": 50,
            "description": "Estoque atual"
          },
          "EST_APOIO": {
            "type": "number",
            "example": 50,
            "description": "Estoque de apoio"
          },
          "PRECO": {
            "type": "number",
            "example": 19.99,
            "description": "Preço 1"
          },
          "PRECO2": {
            "type": "number",
            "example": 17.5,
            "description": "Preço 2"
          }
        },
        "required": [
          "PRO_CODIGO",
          "PRO_CODIGOBAR",
          "PRO_DESCRICAO",
          "MAR_CODIGO",
          "MAR_DESCRICAO",
          "GRU_CODIGO",
          "GRU_DESCRICAO",
          "EST_ATUAL",
          "EST_APOIO",
          "PRECO",
          "PRECO2"
        ]
      },
      "ProductDetailResponseDto": {
        "type": "object",
        "properties": {
          "PRO_CODIGO": {
            "type": "number",
            "example": 101,
            "description": "Código do produto"
          },
          "PRO_CODIGOBAR": {
            "type": "string",
            "example": "7891234567890",
            "description": "Código de barras"
          },
          "PRO_PRCCOMPRA": {
            "type": "number",
            "example": 10,
            "description": "Preço de compra"
          },
          "PRO_PRCCUSTO": {
            "type": "number",
            "example": 12.5,
            "description": "Preço de custo"
          },
          "PRO_PRCCOMPRAFISCAL": {
            "type": "number",
            "example": 11,
            "description": "Preço de compra fiscal"
          },
          "PRO_CUSTOFISCAL": {
            "type": "number",
            "example": 13,
            "description": "Custo fiscal"
          },
          "PRO_PRECO1": {
            "type": "number",
            "example": 19.99,
            "description": "Preço 1"
          }
        },
        "required": [
          "PRO_CODIGO",
          "PRO_CODIGOBAR",
          "PRO_PRCCOMPRA",
          "PRO_PRCCUSTO",
          "PRO_PRCCOMPRAFISCAL",
          "PRO_CUSTOFISCAL",
          "PRO_PRECO1"
        ]
      },
      "ProductBarcodeResponseDto": {
        "type": "object",
        "properties": {
          "PRO_CODIGO": {
            "type": "number",
            "example": 101,
            "description": "Código do produto"
          },
          "PRO_CODIGOBAR": {
            "type": "string",
            "example": "7891234567890",
            "description": "Código de barras"
          },
          "PRO_DESCRICAO": {
            "type": "string",
            "example": "Produto Exemplo",
            "description": "Descrição do produto"
          },
          "MAR_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da marca"
          },
          "MAR_DESCRICAO": {
            "type": "string",
            "example": "Marca Exemplo",
            "description": "Descrição da marca"
          },
          "GRU_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do grupo"
          },
          "GRU_DESCRICAO": {
            "type": "string",
            "example": "Grupo Exemplo",
            "description": "Descrição do grupo"
          },
          "ESTOQUE": {
            "type": "number",
            "example": 50,
            "description": "Estoque total"
          },
          "PRECO": {
            "type": "number",
            "example": 19.99,
            "description": "Preço 1"
          },
          "PRECO2": {
            "type": "number",
            "example": 17.5,
            "description": "Preço 2"
          }
        },
        "required": [
          "PRO_CODIGO",
          "PRO_CODIGOBAR",
          "PRO_DESCRICAO",
          "MAR_CODIGO",
          "MAR_DESCRICAO",
          "GRU_CODIGO",
          "GRU_DESCRICAO",
          "ESTOQUE",
          "PRECO",
          "PRECO2"
        ]
      },
...
      "ServiceProviderResponseDto": {
        "type": "object",
        "properties": {
          "PRE_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do prestador"
          },
          "PRE_NOME": {
            "type": "string",
            "example": "João Silva",
            "description": "Nome do prestador"
          },
          "PRE_APELIDO": {
            "type": "string",
            "example": "Joãozinho",
            "description": "Apelido do prestador"
          },
          "PRE_ENDERECO": {
            "type": "string",
            "example": "Rua Exemplo, 123",
            "description": "Endereço do prestador"
          },
          "PRE_FONE": {
            "type": "string",
            "example": "1144445555",
            "description": "Telefone do prestador"
          },
          "PRE_CELULAR": {
            "type": "string",
            "example": "11999998888",
            "description": "Celular do prestador"
          },
          "PRE_EMAIL": {
            "type": "string",
            "example": "joao@exemplo.com",
            "description": "E-mail do prestador"
          },
          "PRE_SITUACAO": {
            "type": "string",
            "example": "A",
            "description": "Situação do prestador (A = Ativo, I = Inativo)"
          },
          "LOJ_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da loja associada"
          }
        },
        "required": [
          "PRE_CODIGO",
          "PRE_NOME",
          "PRE_APELIDO",
          "PRE_ENDERECO",
          "PRE_FONE",
          "PRE_CELULAR",
          "PRE_EMAIL",
          "PRE_SITUACAO",
          "LOJ_CODIGO"
        ]
      },
      "ServiceProviderDetailResponseDto": {
        "type": "object",
        "properties": {
          "PRE_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do prestador"
          },
          "PRE_NOME": {
            "type": "string",
            "example": "João Silva",
            "description": "Nome do prestador"
          },
          "PRE_APELIDO": {
            "type": "string",
            "example": "Joãozinho",
            "description": "Apelido do prestador"
          },
          "PRE_ENDERECO": {
            "type": "string",
            "example": "Rua Exemplo, 123",
            "description": "Endereço do prestador"
          },
          "PRE_BAIRRO": {
            "type": "string",
            "example": "Centro",
            "description": "Bairro do prestador"
          },
          "PRE_CIDADE": {
            "type": "string",
            "example": "São Paulo",
            "description": "Cidade do prestador"
          },
          "PRE_UF": {
            "type": "string",
            "example": "SP",
            "description": "UF (Estado) do prestador"
          },
          "PRE_CEP": {
            "type": "string",
            "example": "01000-000",
            "description": "CEP do prestador"
          },
          "PRE_FONE": {
            "type": "string",
            "example": "1144445555",
            "description": "Telefone do prestador"
          },
          "PRE_CELULAR": {
            "type": "string",
            "example": "11999998888",
            "description": "Celular do prestador"
          },
          "PRE_EMAIL": {
            "type": "string",
            "example": "joao@exemplo.com",
            "description": "E-mail do prestador"
          },
          "PRE_CPF": {
            "type": "string",
            "example": "123.456.789-00",
            "description": "CPF do prestador"
          },
          "PRE_IDENTIDADE": {
            "type": "string",
            "example": "12.345.678-9",
            "description": "Identidade (RG) do prestador"
          },
          "PRE_SITUACAO": {
            "type": "string",
            "example": "A",
            "description": "Situação do prestador (A = Ativo, I = Inativo)"
          },
          "LOJ_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da loja associada"
          }
        },
        "required": [
          "PRE_CODIGO",
          "PRE_NOME",
          "PRE_APELIDO",
          "PRE_ENDERECO",
          "PRE_BAIRRO",
          "PRE_CIDADE",
          "PRE_UF",
          "PRE_CEP",
          "PRE_FONE",
          "PRE_CELULAR",
          "PRE_EMAIL",
          "PRE_CPF",
          "PRE_IDENTIDADE",
          "PRE_SITUACAO",
          "LOJ_CODIGO"
        ]
      },
...
    }
  }
}